+++ slirp/slirp.c
@@ -1094,7 +1094,7 @@ size_t slirp_socket_can_recv(Slirp *slirp, struct in_addr guest_addr,
 }
 
 void slirp_socket_recv(Slirp *slirp, struct in_addr guest_addr, int guest_port,
-                       const uint8_t *buf, int size)
+                       const uint8_t *buf, size_t size)
 {
     int ret;
     struct socket *so = slirp_find_ctl_socket(slirp, guest_addr, guest_port);
